move getCursorSortDirection function into toggleSort
