+++ src/selectors/index.ts
@@ -25,7 +25,6 @@ export { default as getRootPath } from './getRootPath'
 export { default as getSetting } from './getSetting'
 export { default as getSortPreference } from './getSortPreference'
 export { default as getGlobalSortPreference } from './getGlobalSortPreference'
-export { default as getCursorSortDirection } from './getCursorSortDirection'
 export { default as getStyle } from './getStyle'
 export { default as getThought } from './getThought'
 export { default as getThoughtAfter } from './getThoughtAfter'
